+++ lib/sliver_padding.dart
@@ -5,12 +5,11 @@ class SliverPaddingAll extends StatelessWidget {
   final double padding;
   final Widget sliver;
 
-  SliverPaddingAll(this.padding, {this.sliver});
+  SliverPaddingAll(this.padding, {required this.sliver});
 
   @override
   Widget build(BuildContext context) {
-    return new SliverPadding(
-        padding: new EdgeInsets.all(padding), sliver: sliver);
+    return new SliverPadding(padding: new EdgeInsets.all(padding), sliver: sliver);
   }
 }
 
@@ -23,18 +22,11 @@ class SliverPaddingOnly extends StatelessWidget {
 
   final Widget sliver;
 
-  SliverPaddingOnly({this.left: 0.0,
-    this.top: 0.0,
-    this.right: 0.0,
-    this.bottom: 0.0,
-    this.sliver});
+  SliverPaddingOnly({this.left: 0.0, this.top: 0.0, this.right: 0.0, this.bottom: 0.0, required this.sliver});
 
   @override
   Widget build(BuildContext context) {
-    return new SliverPadding(
-        padding: new EdgeInsets.only(
-            left: left, top: top, right: right, bottom: bottom),
-        sliver: sliver);
+    return new SliverPadding(padding: new EdgeInsets.only(left: left, top: top, right: right, bottom: bottom), sliver: sliver);
   }
 }
 
@@ -43,12 +35,11 @@ class SliverPaddingTop extends StatelessWidget {
   final double padding;
   final Widget sliver;
 
-  SliverPaddingTop(this.padding, {this.sliver});
+  SliverPaddingTop(this.padding, {required this.sliver});
 
   @override
   Widget build(BuildContext context) {
-    return new SliverPadding(
-        padding: new EdgeInsets.only(top: padding), sliver: sliver);
+    return new SliverPadding(padding: new EdgeInsets.only(top: padding), sliver: sliver);
   }
 }
 
@@ -57,12 +48,11 @@ class SliverPaddingBottom extends StatelessWidget {
   final double padding;
   final Widget sliver;
 
-  SliverPaddingBottom(this.padding, {this.sliver});
+  SliverPaddingBottom(this.padding, {required this.sliver});
 
   @override
   Widget build(BuildContext context) {
-    return new SliverPadding(
-        padding: new EdgeInsets.only(bottom: padding), sliver: sliver);
+    return new SliverPadding(padding: new EdgeInsets.only(bottom: padding), sliver: sliver);
   }
 }
 
@@ -71,12 +61,11 @@ class SliverPaddingLeft extends StatelessWidget {
   final double padding;
   final Widget sliver;
 
-  SliverPaddingLeft(this.padding, {this.sliver});
+  SliverPaddingLeft(this.padding, {required this.sliver});
 
   @override
   Widget build(BuildContext context) {
-    return new SliverPadding(
-        padding: new EdgeInsets.only(left: padding), sliver: sliver);
+    return new SliverPadding(padding: new EdgeInsets.only(left: padding), sliver: sliver);
   }
 }
 
@@ -85,12 +74,11 @@ class SliverPaddingRight extends StatelessWidget {
   final double padding;
   final Widget sliver;
 
-  SliverPaddingRight(this.padding, {this.sliver});
+  SliverPaddingRight(this.padding, {required this.sliver});
 
   @override
   Widget build(BuildContext context) {
-    return new SliverPadding(
-        padding: new EdgeInsets.only(right: padding), sliver: sliver);
+    return new SliverPadding(padding: new EdgeInsets.only(right: padding), sliver: sliver);
   }
 }
 
@@ -99,12 +87,11 @@ class SliverPaddingVertical extends StatelessWidget {
   final double padding;
   final Widget sliver;
 
-  SliverPaddingVertical(this.padding, {this.sliver});
+  SliverPaddingVertical(this.padding, {required this.sliver});
 
   @override
   Widget build(BuildContext context) {
-    return new SliverPadding(
-        padding: new EdgeInsets.symmetric(vertical: padding), sliver: sliver);
+    return new SliverPadding(padding: new EdgeInsets.symmetric(vertical: padding), sliver: sliver);
   }
 }
 
@@ -113,25 +100,31 @@ class SliverPaddingHorizontal extends StatelessWidget {
   final double padding;
   final Widget sliver;
 
-  SliverPaddingHorizontal(this.padding, {this.sliver});
+  SliverPaddingHorizontal(this.padding, {required this.sliver});
 
   @override
   Widget build(BuildContext context) {
-    return new SliverPadding(
-        padding: new EdgeInsets.symmetric(horizontal: padding), sliver: sliver);
+    return new SliverPadding(padding: new EdgeInsets.symmetric(horizontal: padding), sliver: sliver);
   }
 }
 
 /// Creates insets with symmetrical horizontal offsets.
 class SliverPaddingSymmetrical extends StatelessWidget {
-  final double padding;
+  final double horizontal;
+  final double vertical;
   final Widget sliver;
 
-  SliverPaddingSymmetrical(this.padding, {this.sliver});
+  SliverPaddingSymmetrical({
+    this.horizontal: 0.0,
+    this.vertical: 0.0,
+    required this.sliver,
+  });
 
   @override
   Widget build(BuildContext context) {
     return new SliverPadding(
-        padding: new EdgeInsets.symmetric(horizontal: padding, vertical: padding), sliver: sliver);
+      padding: new EdgeInsets.symmetric(horizontal: horizontal, vertical: vertical),
+      sliver: sliver,
+    );
   }
 }