+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [1.4.0] - 04/03/21
+
+# Migrated to null-safety
+
+## [1.3.0] - 05/03/20
+
+# SliverPaddingSymmetrical now takes a horizontal and vertical parameter, rather than a single padding parameter
+
 ## [1.2.0] - 19/02/19
 
 # PaddingSymmetrical now takes a horizontal and vertical parameter, rather than a single padding parameter